Refactor VesselSearch so it can be imported in missions actions
