+++ frontend/src/features/VesselSearch/VesselSearchResult.tsx
