+++ frontend/src/features/VesselSearch/VesselSearchResultItem.tsx
