+++ frontend/src/features/VesselSearch/__tests__/removeDuplicatedFoundVessels.test.ts
