+++ frontend/src/features/VesselSearch/utils.ts
