+++ frontend/src/hooks/useClickOutsideWhenOpened.js
@@ -15,7 +15,12 @@ export const useClickOutsideWhenOpened = (ref, isOpened) => {
     // Bind the event listener
     if (isOpened) {
       document.addEventListener('mousedown', handleClickOutside)
+      return
     }
+
+    document.removeEventListener('mousedown', handleClickOutside)
+    setClicked(null)
+
     return () => {
       // Unbind the event listener on clean up
       document.removeEventListener('mousedown', handleClickOutside)
